# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/profiles/verify/send": {
      "post": {
        "description": "사용자 이메일로 인증 코드를 발송합니다.",
        "operationId": "ProfileController_sendVerificationCode",
        "parameters": [],
        "responses": {
          "201": {
            "description": "인증 코드 발송 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "인증 코드가 이메일로 발송되었습니다."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "이메일 인증 코드 발송",
        "tags": [
          "profiles"
        ]
      }
    },
    "/profiles": {
      "post": {
        "description": "새로운 프로필을 생성합니다.",
        "operationId": "ProfileController_createProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfilePayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "프로필 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "홍길동"
                    },
                    "role": {
                      "type": "string",
                      "example": "STUDENT"
                    },
                    "birthDate": {
                      "type": "string",
                      "example": "1990-01-01"
                    },
                    "userId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          },
          "409": {
            "description": "이미 존재하는 프로필"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "프로필 생성",
        "tags": [
          "profiles"
        ]
      },
      "get": {
        "description": "현재 사용자의 모든 프로필을 조회합니다.",
        "operationId": "ProfileController_getMyProfiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "프로필 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "홍길동"
                      },
                      "role": {
                        "type": "string",
                        "example": "STUDENT"
                      },
                      "birthDate": {
                        "type": "string",
                        "example": "1990-01-01"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "내 프로필 목록 조회",
        "tags": [
          "profiles"
        ]
      }
    },
    "/profiles/{id}": {
      "get": {
        "description": "특정 프로필의 상세 정보를 조회합니다.",
        "operationId": "ProfileController_getProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "홍길동"
                    },
                    "role": {
                      "type": "string",
                      "example": "STUDENT"
                    },
                    "birthDate": {
                      "type": "string",
                      "example": "1990-01-01"
                    },
                    "userId": {
                      "type": "number",
                      "example": 1
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "프로필을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "프로필 상세 조회",
        "tags": [
          "profiles"
        ]
      }
    },
    "/profiles/relationships": {
      "post": {
        "description": "학부모와 학생 간의 관계를 등록합니다.",
        "operationId": "ProfileController_registerRelationship",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRelationshipPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "관계 등록 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "관계가 성공적으로 등록되었습니다."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "대상 사용자를 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "관계 등록",
        "tags": [
          "profiles"
        ]
      },
      "get": {
        "description": "학부모는 자녀들의 프로필을, 학생은 부모의 프로필을 조회합니다.",
        "operationId": "ProfileController_getRelatedProfiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "관련 프로필 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "홍길동"
                      },
                      "role": {
                        "type": "string",
                        "example": "STUDENT"
                      },
                      "relationship": {
                        "type": "string",
                        "example": "CHILD"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "권한 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "관련 프로필 조회",
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/permissions/profiles/{profileId}": {
      "get": {
        "description": "특정 프로필의 모든 권한을 조회합니다.",
        "operationId": "PermissionController_getProfilePermissions",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 권한 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "organizationId": {
                        "type": "number",
                        "example": 1
                      },
                      "profileId": {
                        "type": "number",
                        "example": 1
                      },
                      "permission": {
                        "type": "string",
                        "example": "CREATE_COURSE"
                      },
                      "grantedByProfileId": {
                        "type": "number",
                        "example": 1
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-01-01T00:00:00.000Z"
                      },
                      "expiresAt": {
                        "type": "string",
                        "example": "2024-12-31T00:00:00.000Z"
                      },
                      "organization": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "테스트 학교"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "프로필 권한 목록 조회",
        "tags": [
          "permissions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Thinking API Server",
    "description": "수업 관리 시스템 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT 토큰을 입력하세요",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateProfilePayload": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "프로필 이름",
            "example": "홍길동"
          },
          "role": {
            "type": "string",
            "description": "사용자 역할",
            "enum": [
              "STUDENT",
              "PARENT",
              "TEACHER",
              "ADMIN"
            ],
            "example": "STUDENT"
          },
          "birthDate": {
            "type": "string",
            "description": "생년월일 (YYYY-MM-DD 형식)",
            "example": "1990-01-01"
          },
          "verificationCode": {
            "type": "string",
            "description": "이메일 인증 코드",
            "example": "123456"
          }
        },
        "required": [
          "name",
          "role",
          "birthDate",
          "verificationCode"
        ]
      },
      "RegisterRelationshipPayload": {
        "type": "object",
        "properties": {
          "targetUserEmail": {
            "type": "string",
            "description": "관계를 등록할 대상 사용자의 이메일",
            "example": "student@example.com"
          }
        },
        "required": [
          "targetUserEmail"
        ]
      },
...
    }
  }
}